+++ src/main/ml-modules/root/ds/lux/document/create.mjs
@@ -1,5 +1,5 @@
 import { handleRequest } from '../../../lib/securityLib.mjs';
-import { createSet } from '../../../lib/setLib.mjs';
+import { createDocument } from '../../../lib/crudLib.mjs';
 
 const unitName = external.unitName;
 const docNode = external.doc; // Do not use getObjectFromNode
@@ -7,5 +7,5 @@ const lang = external.lang;
 
 handleRequest(function () {
   declareUpdate();
-  return createSet(docNode, lang);
+  return createDocument(docNode, lang);
 }, unitName);
