+++ src/main/ml-modules/root/ds/lux/document/read.mjs
@@ -1,10 +1,10 @@
 import { handleRequest } from '../../../lib/securityLib.mjs';
-import { get } from '../../../lib/crudLib.mjs';
+import { readDocument } from '../../../lib/crudLib.mjs';
 
 const unitName = external.unitName;
 const uri = external.uri;
 const profile = external.profile;
 const lang = external.lang;
 handleRequest(function () {
-  return get(uri, profile, lang);
+  return readDocument(uri, profile, lang);
 }, unitName);
