+++ src/main/ml-modules/root/ds/lux/document/update.mjs
@@ -1,5 +1,5 @@
 import { handleRequest } from '../../../lib/securityLib.mjs';
-import { updateSet } from '../../../lib/setLib.mjs';
+import { updateDocument } from '../../../lib/crudLib.mjs';
 
 const unitName = external.unitName;
 const docNode = external.doc; // Do not use getObjectFromNode
@@ -7,5 +7,5 @@ const lang = external.lang;
 
 handleRequest(function () {
   declareUpdate();
-  return updateSet(docNode, lang);
+  return updateDocument(docNode, lang);
 }, unitName);
